+++ assets/2025-04-06_No_Adjacent_Characters/rust/src/a5_multithreading_a2.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use rayon::prelude::*;
 
-pub fn run(input: &mut [u128]) {
+pub fn run(input: &mut [u8]) {
     if input.len() <= 1 {
         return;
     }
@@ -23,7 +23,7 @@ pub fn run(input: &mut [u128]) {
             acc
         });
 
-    let mut elements: Vec<(u128, usize)> = counts.into_iter().collect();
+    let mut elements: Vec<(u8, usize)> = counts.into_iter().collect();
     elements.sort_by(|a, b| b.1.cmp(&a.1));
 
     let max_count = elements[0].1;
